Daemon should support reload config
This change adds support for changing the limit configuration of a running sphinxd.
It does this by listening for the HUP signal, reloading the config file when
that happens, and changing the ratelimiter object.

We aren't worrying about pre-existing buckets that get affected when we reload,
they will expire out.

I also tested this manually with Apache Bench